Renamed ee-client to queen-elizabeth
