+++ queen-elizabeth/config/.gitignore
