+++ queen-elizabeth/src/main/java/gov/va/api/health/queenelizabeth/controller/ErrorResponse.java
@@ -1,4 +1,4 @@
-package gov.va.api.health.eeclient.controller;
+package gov.va.api.health.queenelizabeth.controller;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import javax.xml.bind.annotation.XmlAccessType;
