+++ queen-elizabeth/src/main/resources/application.properties
@@ -5,6 +5,8 @@ ee.endpoint.url=unset
 ee.header.username=unset
 ee.header.password=unset
 ee.request.name=unset
+ee.truststore.path=unset
+ee.truststore.password=unset
 
 #SSL Properties
 security.require-ssl=true
