+++ queen-elizabeth/src/main/resources/ee-api-v1.yaml
